requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==4.0.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
imbalanced-learn==0.9.0
imblearn==0.0
joblib==1.1.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
numpy==1.22.1
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.4.0
Pillow==9.0.1
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
